proj1/requirements.txt

accelerate==1.10.1
arrow==1.3.0
bitsandbytes==0.47.0
datasets==3.6.0
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl
huggingface-hub==0.34.4
matplotlib==3.10.5
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
peft==0.17.1
pillow==11.3.0
pyarrow==21.0.0
regex==2025.8.29
sacremoses==0.1.1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.1
tokenizers==0.22.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
wandb==0.21.0

